+++ pages/services/ServicesBreadcam.js
@@ -1,14 +1,14 @@
 
 
-const ServicesBreadcam = () => {
+const ServicesBreadcam = ({ logoText }) => {
   return (
     <section className="servicesBreadcam">
 
-        <div className="content h-full flex justify-center items-center text-white pt-[13%] ">
+      <div className="content h-full flex justify-center items-center text-white pt-[11%] ">
         <h1 className="text-3xl h-full text-white my-5 text-center font-semibold mb-5">
-        OUR SERVICES
-      </h1>
-        </div>
+          {logoText}
+        </h1>
+      </div>
 
     </section>
   )
