+++ styles/home.css
@@ -231,7 +231,7 @@
 .servicesBreadcam{
   background: url('https://ihsan.templines.org/wp-content/uploads/2019/10/services-head-1.png');
   width: 100%;
-  height: 450px;
+  height: 350px;
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
